Move BlendLatentsInvocation to its own file. No functional changes.
